requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
altair==5.3.0
altgraph==0.17.4
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
auto-py-to-exe==2.44.1
awsebcli==3.20.10
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bitsets==0.8.4
bitstring==3.1.9
bleach==6.0.0
blinker==1.7.0
botocore==1.31.85
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.0
cement==2.8.2
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
comm==0.1.2
concepts==0.9.2
contourpy==1.0.7
cryptography==38.0.3
cvzone==1.5.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
ecdsa==0.18.0
Eel==0.16.0
esptool==4.3
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.92.0
fastjsonschema==2.16.2
features==0.5.12
fileconfig==0.6.1
filelock==3.9.0
Flask==3.0.3
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.10.0
future==1.0.0
gast==0.4.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-speech==2.17.3
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.12.1
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.30
jaxlib==0.4.30
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_fileid==0.7.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
kaggle==1.5.12
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mechanism==1.1.7
mistune==2.0.5
ml-dtypes==0.4.0
mpmath==1.3.0
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.22.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pydeck==0.9.1
Pygments==2.14.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-json-logger==2.0.6
python-multipart==0.0.5
python-slugify==8.0.0
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
reedsolo==1.5.4
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.13.2
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
split-folders==0.5.1
stack-data==0.6.2
starlette==0.25.0
streamlit==1.37.0
sympy==1.12
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-serving-api==2.5.0
termcolor==1.1.0
terminado==0.17.1
text-unidecode==1.3
tflearn==0.5.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2024.1
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.20.0
watchdog==4.0.1
wcwidth==0.1.9
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.1
whichcraft==0.6.1
wrapt==1.14.1
y-py==0.5.5
ypy-websocket==0.8.2
zope.event==5.0
zope.interface==6.4.post2
